requirements.txt
flask
requests
python-dotenv
llama_cpp_python
pycryptodome
schedule
dnspython
beautifulsoup4
openai
